+++ src/engine/magnetization.go
@@ -10,8 +10,6 @@ import (
 
 var M magnetization // reduced magnetization (unit length)
 
-func init() { DeclLValue("m", &M, `Reduced magnetization (unit length)`) }
-
 // Special buffered quantity to store magnetization
 // makes sure it's normalized etc.
 type magnetization struct {
