+++ src/engine/run.go
@@ -35,19 +35,7 @@ var (
 )
 
 func init() {
-	DeclFunc("Run", Run, "Run the simulation for a time in seconds")
-	DeclFunc("RunWithoutPrecession", RunWithoutPrecession, "Run the simulation for a time in seconds with precession disabled")
-	DeclFunc("Steps", Steps, "Run the simulation for a number of time steps")
-	DeclFunc("RunWhile", RunWhile, "Run while condition function is true")
-	DeclFunc("SetSolver", SetSolver, "Set solver type. 1:Euler, 2:Heun, 3:Bogaki-Shampine, 4: Runge-Kutta (RK45), 5: Dormand-Prince, 6: Fehlberg, -1: Backward Euler")
-	DeclTVar("t", &Time, "Total simulated time (s)")
-	DeclVar("step", &NSteps, "Total number of time steps taken")
-	DeclVar("MinDt", &MinDt, "Minimum time step the solver can take (s)")
-	DeclVar("MaxDt", &MaxDt, "Maximum time step the solver can take (s)")
-	DeclVar("MaxErr", &MaxErr, "Maximum error per step the solver can tolerate (default = 1e-5)")
-	DeclVar("Headroom", &Headroom, "Solver headroom (default = 0.8)")
-	DeclVar("FixDt", &FixDt, "Set a fixed time step, 0 disables fixed step (which is the default)")
-	DeclFunc("Exit", Exit, "Exit from the program")
+
 	SetSolver(DORMANDPRINCE)
 	_ = NewScalarValue("dt", "s", "Time Step", func() float64 { return Dt_si })
 	_ = NewScalarValue("LastErr", "", "Error of last step", func() float64 { return LastErr })
